playlist-songs

+++ src/services/postgres/PlaylistSongsService.js
@@ -1,14 +1,16 @@
 const { Pool } = require('pg');
 const { nanoid } = require('nanoid');
 const InvariantError = require('../../exceptions/InvariantError');
-const NotFoundError = require('../../exceptions/NotFoundError');
+// const NotFoundError = require('../../exceptions/NotFoundError');
 
 class PlaylistSongsService {
-  constructor() {
+  constructor(playlistsService) {
     this._pool = new Pool();
+    this._playlistsService = playlistsService;
   }
 
-  async addPlaylistSong(playlistId, songId) {
+  async addPlaylistSong(playlistId, songId, userId) {
+    await this._playlistsService.verifyPlaylistAccess(playlistId, userId);
     const id = `plsong-${nanoid(16)}`;
 
     const query = {
@@ -17,16 +19,19 @@ class PlaylistSongsService {
     };
 
     const result = await this._pool.query(query);
-    // console.log(`addSongPlaylist: ${result}`);
+    // const resultString = JSON.stringify(result);
+    // console.log(`addSongPlaylist: ${resultString}`);
 
-    if (!result.rows[0].id) {
+    if (!result.rowCount) {
       throw new InvariantError('Lagu gagal ditambahkan ke dalam playlist');
     }
 
-    return result.rows[0].id;
+    // return result.rows[0].id;
   }
 
-  async getAllPlaylistSongs(playlistId) {
+  async getAllPlaylistSongs(playlistId, userId) {
+    await this._playlistsService.verifyPlaylistAccess(playlistId, userId);
+    const playlist = await this._playlistsService.getPlaylistById(userId);
     const query = {
       text: `SELECT songs.id, songs.title, songs.performer
       FROM songs LEFT JOIN
@@ -36,20 +41,22 @@ class PlaylistSongsService {
     };
 
     const result = await this._pool.query(query);
-    // console.log(result);
+    // console.log(`getAllPlaylistSong: ${result}`);
     if (!result.rows.length) {
-      throw new NotFoundError('Lagu tidak ditemukan');
+      throw new InvariantError('Gagal menemukan lagu');
     }
-    return result.rows;
+    return { ...playlist, songs: result.rows };
   }
 
-  async deletePlaylistSong(playlistId, songId) {
+  async deletePlaylistSong(playlistId, songId, userId) {
+    await this._playlistsService.verifyPlaylistAccess(playlistId, userId);
     const query = {
       text: 'DELETE FROM playlist_songs WHERE playlist_id = $1 AND song_id = $2 RETURNING id',
       values: [playlistId, songId],
     };
 
     const result = await this._pool.query(query);
+    // console.log(`deleteAllPlaylistSong: ${result}`);
 
     if (!result.rows.length) {
       throw new InvariantError('Lagu gagal dihapus');
